Split TagConfig god-file: extract 7 standalone types into 3 files
Relocate complete top-level types out of the 7,877-line TagConfig.cs
(now 6,427) into cohesive same-namespace files — pure file moves, no
semantic change:

- Core/TaggingModels.cs    — DisciplineProfile, TaggingStats,
                             ValidationErrorType, ValidationError
- Core/ISO19650Validator.cs — ISO19650Validator
- Core/TagIntelligence.cs   — TagIntelligence, StingPluginHooks

Each new file carries the namespace + required usings. Verified: every
moved type resolves to exactly one definition, BuildAndWriteTag/
WriteTag7All/TagIsComplete remain in TagConfig.cs, and all files are
brace-balanced. Needs a Windows dotnet build to confirm.
