+++ StingTools/Core/TagConfig.cs
@@ -29,926 +29,7 @@ namespace StingTools.Core
     // SeqScheme enum relocated to Core/SeqAssigner.cs (same namespace) alongside
     // the pure sequence-assignment logic it parameterises.
 
-    /// <summary>
-    /// GAP-FIX: Per-discipline tagging profile. Allows each discipline to have different
-    /// collision handling, SEQ scheme, and default token values.
-    /// Loaded from DISCIPLINE_PROFILES in project_config.json.
-    /// </summary>
-    public class DisciplineProfile
-    {
-        // ── v1 properties (collision/SEQ/token defaults) ──
 
-        /// <summary>Collision mode override for this discipline (Skip/Overwrite/AutoIncrement). Null = use global.</summary>
-        public TagCollisionMode? CollisionMode { get; set; }
-
-        /// <summary>SEQ scheme override (Numeric/Alpha/ZonePrefix/DiscPrefix). Null = use global.</summary>
-        public SeqScheme? SeqScheme { get; set; }
-
-        /// <summary>Default ZONE code for this discipline. Null = use auto-detect.</summary>
-        public string DefaultZone { get; set; }
-
-        /// <summary>Default LOC code for this discipline. Null = use auto-detect.</summary>
-        public string DefaultLoc { get; set; }
-
-        /// <summary>Whether to include zone in SEQ key for this discipline. Null = use global.</summary>
-        public bool? SeqIncludeZone { get; set; }
-
-        /// <summary>Custom SEQ pad width for this discipline (e.g., 3 for 001, 5 for 00001). Null = use global.</summary>
-        public int? SeqPadWidth { get; set; }
-
-        // ── v2 properties (validation constraints) ──
-
-        /// <summary>Default DISC code for this profile (e.g., "M").</summary>
-        public string DefaultDisc { get; set; }
-
-        /// <summary>Allowed SYS codes for this discipline. Empty set means no restriction.</summary>
-        public HashSet<string> AllowedSysCodes { get; set; } = new HashSet<string>(StringComparer.OrdinalIgnoreCase);
-
-        /// <summary>Allowed FUNC codes for this discipline. Empty set means no restriction.</summary>
-        public HashSet<string> AllowedFuncCodes { get; set; } = new HashSet<string>(StringComparer.OrdinalIgnoreCase);
-
-        /// <summary>Default PROD code when family-aware detection yields a generic result.</summary>
-        public string DefaultProd { get; set; }
-
-        /// <summary>Default STATUS value for this discipline.</summary>
-        public string DefaultStatus { get; set; }
-
-
-        /// <summary>When true, SYS/FUNC must be in AllowedSysCodes/AllowedFuncCodes.</summary>
-        public bool ValidationStrictness { get; set; }
-
-        /// <summary>Tokens that must be non-empty for compliant tags (e.g., ["DISC","SYS","FUNC","PROD","SEQ"]).</summary>
-        public HashSet<string> RequiredTokens { get; set; } = new HashSet<string>(StringComparer.OrdinalIgnoreCase);
-
-        /// <summary>
-        /// Default paragraph depth (1-10) for this discipline's TAG7 tier
-        /// visibility. Null = use the global ParaDepth value. Used by
-        /// SetParagraphDepthCommand's "By discipline" scope and read by
-        /// WriteTag7All when no per-element override is set.
-        /// </summary>
-        public int? DefaultParagraphDepth { get; set; }
-
-        /// <summary>Parse a DisciplineProfile from a JSON dictionary.</summary>
-        public static DisciplineProfile FromDict(Dictionary<string, object> dict)
-        {
-            var p = new DisciplineProfile();
-            if (dict == null) return p;
-
-            if (dict.TryGetValue("collision_mode", out object cm) && cm is string cms)
-            {
-                if (Enum.TryParse<TagCollisionMode>(cms, true, out var parsed)) p.CollisionMode = parsed;
-            }
-            if (dict.TryGetValue("seq_scheme", out object ss) && ss is string sss)
-            {
-                if (Enum.TryParse<SeqScheme>(sss, true, out var parsed)) p.SeqScheme = parsed;
-            }
-            if (dict.TryGetValue("default_zone", out object dz) && dz is string dzs && !string.IsNullOrWhiteSpace(dzs))
-                p.DefaultZone = dzs;
-            if (dict.TryGetValue("default_loc", out object dl) && dl is string dls && !string.IsNullOrWhiteSpace(dls))
-                p.DefaultLoc = dls;
-            if (dict.TryGetValue("default_status", out object ds) && ds is string dss && !string.IsNullOrWhiteSpace(dss))
-                p.DefaultStatus = dss;
-            if (dict.TryGetValue("seq_include_zone", out object siz))
-            {
-                if (siz is bool b) p.SeqIncludeZone = b;
-                else if (siz is string sizs) p.SeqIncludeZone = sizs.Equals("true", StringComparison.OrdinalIgnoreCase);
-            }
-            if (dict.TryGetValue("seq_pad_width", out object spw))
-            {
-                if (spw is long l) p.SeqPadWidth = (int)l;
-                else if (int.TryParse(spw?.ToString(), out int iv)) p.SeqPadWidth = iv;
-            }
-            // v2 properties from dict
-            if (dict.TryGetValue("default_disc", out object dd) && dd is string dds && !string.IsNullOrWhiteSpace(dds))
-                p.DefaultDisc = dds;
-            if (dict.TryGetValue("default_prod", out object dprod) && dprod is string dprods && !string.IsNullOrWhiteSpace(dprods))
-                p.DefaultProd = dprods;
-            if (dict.TryGetValue("validation_strictness", out object vs))
-            {
-                if (vs is bool vsb) p.ValidationStrictness = vsb;
-                else if (vs is string vss) p.ValidationStrictness = vss.Equals("true", StringComparison.OrdinalIgnoreCase);
-            }
-            if (dict.TryGetValue("default_paragraph_depth", out object dpd))
-            {
-                if (dpd is long dpdl && dpdl >= 1 && dpdl <= 10) p.DefaultParagraphDepth = (int)dpdl;
-                else if (int.TryParse(dpd?.ToString(), out int dpdi) && dpdi >= 1 && dpdi <= 10)
-                    p.DefaultParagraphDepth = dpdi;
-            }
-            return p;
-        }
-    }
-
-    /// <summary>
-    /// Tracks tagging operation statistics across a batch for rich post-operation reporting.
-    /// Captures per-category counts, collision details, skipped elements, warnings, and
-    /// discipline/system/level breakdowns. Thread-safe for single-transaction use.
-    /// </summary>
-    public class TaggingStats
-    {
-        public int TotalTagged { get; private set; }
-        public int TotalSkipped { get; private set; }
-        public int TotalOverwritten { get; private set; }
-        public int TotalCollisions { get; private set; }
-        public int MaxCollisionDepth { get; private set; }
-        public readonly Dictionary<string, int> TaggedByCategory = new Dictionary<string, int>();
-        public readonly Dictionary<string, int> TaggedByDisc = new Dictionary<string, int>();
-        public readonly Dictionary<string, int> TaggedBySys = new Dictionary<string, int>();
-        public readonly Dictionary<string, int> TaggedByLevel = new Dictionary<string, int>();
-        public readonly Dictionary<string, int> SkippedByCategory = new Dictionary<string, int>();
-        public readonly Dictionary<string, int> OverwrittenByCategory = new Dictionary<string, int>();
-        public readonly Dictionary<string, int> OverwrittenByDisc = new Dictionary<string, int>();
-        public readonly Dictionary<string, int> OverwrittenBySys = new Dictionary<string, int>();
-        public readonly Dictionary<string, int> OverwrittenByLevel = new Dictionary<string, int>();
-        public readonly List<string> Warnings = new List<string>();
-        public readonly List<(string tag, int depth)> CollisionDetails = new List<(string, int)>();
-
-        /// <summary>PERF-02: Inline count of elements with empty FUNC after pipeline.</summary>
-        public int EmptyFuncCount { get; private set; }
-        /// <summary>PERF-02: Inline count of elements with empty PROD after pipeline.</summary>
-        public int EmptyProdCount { get; private set; }
-        /// <summary>PERF-R13: Count of elements that defaulted to LOC=BLD1 (throttled from per-element warnings).</summary>
-        public int DefaultLocCount { get; set; }
-        /// <summary>PERF-R13: Count of elements that defaulted to ZONE=Z01 (throttled from per-element warnings).</summary>
-        public int DefaultZoneCount { get; set; }
-
-        /// <summary>PERF-02: Track empty FUNC/PROD inline during tagging loop to avoid post-loop re-scan.</summary>
-        public void RecordEmptyTokens(string func, string prod)
-        {
-            if (string.IsNullOrEmpty(func)) EmptyFuncCount++;
-            if (string.IsNullOrEmpty(prod)) EmptyProdCount++;
-        }
-
-        public void RecordTagged(string category, string disc, string sys, string lvl)
-        {
-            TotalTagged++;
-            Increment(TaggedByCategory, category);
-            if (!string.IsNullOrEmpty(disc)) Increment(TaggedByDisc, disc);
-            if (!string.IsNullOrEmpty(sys)) Increment(TaggedBySys, sys);
-            if (!string.IsNullOrEmpty(lvl)) Increment(TaggedByLevel, lvl);
-        }
-
-        public void RecordSkipped(string category)
-        {
-            TotalSkipped++;
-            Increment(SkippedByCategory, category);
-        }
-
-        public void RecordOverwritten(string category, string disc = null, string sys = null, string lvl = null)
-        {
-            TotalOverwritten++;
-            Increment(OverwrittenByCategory, category);
-            if (!string.IsNullOrEmpty(disc)) Increment(OverwrittenByDisc, disc);
-            if (!string.IsNullOrEmpty(sys)) Increment(OverwrittenBySys, sys);
-            if (!string.IsNullOrEmpty(lvl)) Increment(OverwrittenByLevel, lvl);
-        }
-
-        public void RecordCollision(string tag, int depth)
-        {
-            TotalCollisions++;
-            if (depth > MaxCollisionDepth) MaxCollisionDepth = depth;
-            // Keep the top 20 collisions by depth (deepest = most concerning)
-            if (CollisionDetails.Count < 20)
-            {
-                CollisionDetails.Add((tag, depth));
-            }
-            else
-            {
-                // Replace the shallowest collision if this one is deeper
-                int minIdx = 0;
-                for (int i = 1; i < CollisionDetails.Count; i++)
-                    if (CollisionDetails[i].depth < CollisionDetails[minIdx].depth)
-                        minIdx = i;
-                if (depth > CollisionDetails[minIdx].depth)
-                    CollisionDetails[minIdx] = (tag, depth);
-            }
-        }
-
-        public void RecordWarning(string warning)
-        {
-            if (Warnings.Count < 100)
-                Warnings.Add(warning);
-        }
-
-        /// <summary>Build a multi-line report string for TaskDialog display.</summary>
-        public string BuildReport()
-        {
-            var sb = new System.Text.StringBuilder();
-            sb.AppendLine($"  Tagged:       {TotalTagged:N0}");
-            sb.AppendLine($"  Skipped:      {TotalSkipped:N0}");
-            if (TotalOverwritten > 0)
-                sb.AppendLine($"  Overwritten:  {TotalOverwritten:N0}");
-            if (TotalCollisions > 0)
-            {
-                sb.AppendLine($"  Collisions:   {TotalCollisions:N0} resolved (max depth: {MaxCollisionDepth})");
-                // Show top 5 deepest collisions (most concerning)
-                foreach (var (tag, depth) in CollisionDetails.OrderByDescending(c => c.depth).Take(5))
-                    sb.AppendLine($"    • {tag} (bumped {depth}×)");
-                if (TotalCollisions > 5)
-                    sb.AppendLine($"    ... and {TotalCollisions - 5} more");
-            }
-
-            if (TaggedByDisc.Count > 0)
-            {
-                sb.AppendLine();
-                sb.AppendLine("  By Discipline:");
-                foreach (var kvp in TaggedByDisc.OrderByDescending(x => x.Value))
-                    sb.AppendLine($"    {kvp.Key,-6} {kvp.Value,5}");
-            }
-            if (TaggedBySys.Count > 0)
-            {
-                sb.AppendLine();
-                sb.AppendLine("  By System:");
-                foreach (var kvp in TaggedBySys.OrderByDescending(x => x.Value).Take(8))
-                    sb.AppendLine($"    {kvp.Key,-8} {kvp.Value,5}");
-            }
-            if (TaggedByLevel.Count > 0)
-            {
-                sb.AppendLine();
-                sb.AppendLine("  By Level:");
-                foreach (var kvp in TaggedByLevel.OrderBy(x => x.Key))
-                    sb.AppendLine($"    {kvp.Key,-6} {kvp.Value,5}");
-            }
-            if (TotalOverwritten > 0 && OverwrittenByDisc.Count > 0)
-            {
-                sb.AppendLine();
-                sb.AppendLine("  Overwritten By Discipline:");
-                foreach (var kvp in OverwrittenByDisc.OrderByDescending(x => x.Value))
-                    sb.AppendLine($"    {kvp.Key,-6} {kvp.Value,5}");
-            }
-            if (Warnings.Count > 0)
-            {
-                sb.AppendLine();
-                sb.AppendLine($"  Warnings ({Warnings.Count}):");
-                foreach (string w in Warnings.Take(10))
-                    sb.AppendLine($"    ⚠ {w}");
-                if (Warnings.Count > 10)
-                    sb.AppendLine($"    ... and {Warnings.Count - 10} more");
-            }
-
-            return sb.ToString();
-        }
-
-        private static void Increment(Dictionary<string, int> dict, string key)
-        {
-            if (string.IsNullOrEmpty(key)) return;
-            dict.TryGetValue(key, out int count);
-            dict[key] = count + 1;
-        }
-    }
-
-    /// <summary>Categorizes ISO 19650 validation errors for separate counting.</summary>
-    public enum ValidationErrorType
-    {
-        /// <summary>Token value does not match allowed code list (e.g., DISC 'X' not valid).</summary>
-        TokenFormat,
-        /// <summary>Token value is empty/missing.</summary>
-        TokenEmpty,
-        /// <summary>Cross-validation mismatch between related tokens or element category.</summary>
-        CrossValidation
-    }
-
-    /// <summary>Structured validation error with message and categorized type.</summary>
-    public class ValidationError
-    {
-        public string Message { get; set; }
-        public ValidationErrorType Type { get; set; }
-
-        public ValidationError(string message, ValidationErrorType type)
-        {
-            Message = message;
-            Type = type;
-        }
-
-        public override string ToString() => Message;
-    }
-
-    /// <summary>
-    /// ISO 19650 naming and coding validation. Enforces that all token values
-    /// conform to the allowed code lists defined in the tag configuration.
-    /// </summary>
-    public static class ISO19650Validator
-    {
-        /// <summary>FLEX-001: Custom DISC codes from project_config.json CUSTOM_VALID_DISC.
-        /// When non-empty, these are ADDED to the built-in codes for validation.</summary>
-        public static HashSet<string> CustomDiscCodes { get; internal set; } = new HashSet<string>(StringComparer.OrdinalIgnoreCase);
-        /// <summary>FLEX-001: Custom SYS codes from project_config.json CUSTOM_VALID_SYS.</summary>
-        public static HashSet<string> CustomSysCodes { get; internal set; } = new HashSet<string>(StringComparer.OrdinalIgnoreCase);
-        /// <summary>FLEX-001: Custom FUNC codes from project_config.json CUSTOM_VALID_FUNC.</summary>
-        public static HashSet<string> CustomFuncCodes { get; internal set; } = new HashSet<string>(StringComparer.OrdinalIgnoreCase);
-        /// <summary>FLEX-001: Custom LOC codes from project_config.json CUSTOM_VALID_LOC.</summary>
-        public static HashSet<string> CustomLocCodes { get; internal set; } = new HashSet<string>(StringComparer.OrdinalIgnoreCase);
-        /// <summary>FLEX-001: Custom ZONE codes from project_config.json CUSTOM_VALID_ZONE.</summary>
-        public static HashSet<string> CustomZoneCodes { get; internal set; } = new HashSet<string>(StringComparer.OrdinalIgnoreCase);
-
-        /// <summary>Built-in valid discipline codes per ISO 19650.</summary>
-        private static readonly HashSet<string> _builtInDiscCodes = new HashSet<string>(StringComparer.OrdinalIgnoreCase)
-        {
-            "M", "E", "P", "A", "S", "FP", "LV", "G"
-        };
-
-        /// <summary>Valid discipline codes: built-in + custom from config (FLEX-001). Cached to avoid per-access allocation.</summary>
-        private static HashSet<string> _cachedValidDiscCodes;
-        public static HashSet<string> ValidDiscCodes
-        {
-            get
-            {
-                if (CustomDiscCodes.Count == 0) return _builtInDiscCodes;
-                var cached = _cachedValidDiscCodes;
-                if (cached != null) return cached;
-                var combined = new HashSet<string>(_builtInDiscCodes, StringComparer.OrdinalIgnoreCase);
-                foreach (string c in CustomDiscCodes) combined.Add(c);
-                _cachedValidDiscCodes = combined;
-                return combined;
-            }
-        }
-
-        /// <summary>
-        /// Valid system codes per CIBSE / Uniclass 2015 / ISO 19650.
-        /// DCW = Domestic Cold Water (Uniclass Ss_55_70_38_15, CAWS S10)
-        /// DHW = Domestic Hot Water (Uniclass Ss_55_70_38, CAWS S11)
-        /// HWS = Heating Water System (LTHW heating circuits)
-        /// SAN = Sanitary drainage (Uniclass Ss_50_30_04, CAWS R11)
-        /// GAS = Gas Supply (Uniclass Ss_55_20_34, CAWS S63)
-        /// RWD = Rainwater Drainage (Uniclass Ss_50_30_02, CAWS R10)
-        /// CSV element type codes (P-WSP, P-DRN) map to these system codes.
-        /// </summary>
-        /// <summary>ISO 19650 fallback SYS codes used when no project-specific config is loaded.</summary>
-        private static readonly HashSet<string> _fallbackSysCodes = new HashSet<string>(StringComparer.OrdinalIgnoreCase)
-        {
-            "HVAC", "HWS", "DHW", "DCW", "SAN", "RWD", "GAS", "FP", "LV",
-            "FLS", "COM", "ICT", "NCL", "SEC",
-            "ARC", "STR", "GEN"
-        };
-
-        /// <summary>
-        /// Valid system codes. Dynamically derived from TagConfig.SysMap keys when available,
-        /// falling back to hardcoded CIBSE/Uniclass codes when SysMap is not loaded.
-        /// This ensures custom project SYS codes in project_config.json are accepted.
-        /// BUG-06: Uses Count > 0 check to handle empty but non-null SysMap.
-        /// </summary>
-        /// <summary>PERF-01: Cached SYS codes set — rebuilt only when SysMap changes (via InvalidateValidatorCaches).</summary>
-        private static HashSet<string> _cachedValidSysCodes;
-        public static HashSet<string> ValidSysCodes
-        {
-            get
-            {
-                if (TagConfig.SysMap == null || TagConfig.SysMap.Count == 0) return _fallbackSysCodes;
-                var cached = _cachedValidSysCodes;
-                if (cached != null) return cached;
-                var set = new HashSet<string>(TagConfig.SysMap.Keys, StringComparer.OrdinalIgnoreCase);
-                if (CustomSysCodes.Count > 0) foreach (string c in CustomSysCodes) set.Add(c);
-                _cachedValidSysCodes = set;
-                return set;
-            }
-        }
-
-        /// <summary>ISO 19650 fallback FUNC codes used when no project-specific config is loaded.</summary>
-        private static readonly HashSet<string> _fallbackFuncCodes = new HashSet<string>(StringComparer.OrdinalIgnoreCase)
-        {
-            "SUP", "HTG", "DCW", "SAN", "RWD", "GAS", "FP", "PWR", "FLS",
-            "COM", "ICT", "NCL", "SEC",
-            "FIT", "STR", "GEN",
-            "EXH", "RTN", "FRA", "DHW"
-        };
-
-        /// <summary>
-        /// BUG-06: Valid FUNC codes derived from TagConfig.FuncMap keys when available,
-        /// so project-specific codes added via project_config.json are accepted.
-        /// Falls back to ISO 19650 standard codes when FuncMap is empty.
-        /// </summary>
-        /// <summary>PERF-01: Cached FUNC codes set — rebuilt only when FuncMap changes (via InvalidateValidatorCaches).</summary>
-        private static HashSet<string> _cachedValidFuncCodes;
-        public static HashSet<string> ValidFuncCodes
-        {
-            get
-            {
-                if (TagConfig.FuncMap == null || TagConfig.FuncMap.Count == 0) return _fallbackFuncCodes;
-                var cached = _cachedValidFuncCodes;
-                if (cached != null) return cached;
-                var set = new HashSet<string>(TagConfig.FuncMap.Keys, StringComparer.OrdinalIgnoreCase);
-                if (CustomFuncCodes.Count > 0) foreach (string c in CustomFuncCodes) set.Add(c);
-                _cachedValidFuncCodes = set;
-                return set;
-            }
-        }
-
-        // F-08: Cached HashSet<string> for LocCodes and ZoneCodes — avoids O(n) List.Contains in ValidateToken
-        private static HashSet<string> _locCodesSet;
-        private static HashSet<string> _zoneCodesSet;
-
-        /// <summary>PERF-01: Invalidate cached validator code sets. Call after LoadFromFile/LoadDefaults or custom code changes.</summary>
-        internal static void InvalidateValidatorCaches()
-        {
-            _cachedValidDiscCodes = null;
-            _cachedValidSysCodes = null;
-            _cachedValidFuncCodes = null;
-            _locCodesSet = null;  // F-08
-            _zoneCodesSet = null; // F-08
-            _tokenValidationCache.Clear(); // Phase 78: Clear memoized validation results
-            _validFuncsCsvLoaded = false;  // Force reload of CSV-derived func codes
-            EnsureValidFuncsLoaded();
-        }
-
-        /// <summary>
-        /// Validate a single token value against its allowed code list.
-        /// Returns null if valid, or an error message string if invalid.
-        /// </summary>
-        public static string ValidateToken(string tokenName, string value)
-        {
-            if (string.IsNullOrEmpty(value))
-                return $"{tokenName}: empty value";
-
-            // Use if/else chain instead of switch — ParamRegistry properties are not compile-time constants
-            if (tokenName == ParamRegistry.DISC)
-            {
-                if (!ValidDiscCodes.Contains(value))
-                    return $"DISC '{value}' not in valid set ({string.Join(",", ValidDiscCodes)})";
-            }
-            else if (tokenName == ParamRegistry.LOC)
-            {
-                // FLEX-001: Accept custom LOC codes from config before strict/lenient check
-                if (CustomLocCodes.Count > 0 && CustomLocCodes.Contains(value))
-                    return null; // Custom code accepted
-                if (TagConfig.ValidateStrictMode)
-                {
-                    // F-08: Use cached HashSet for O(1) lookup instead of List.Contains O(n)
-                    var locSet = _locCodesSet ??= new HashSet<string>(TagConfig.LocCodes, StringComparer.OrdinalIgnoreCase);
-                    if (!locSet.Contains(value))
-                        return $"LOC '{value}' not in valid set ({string.Join(",", TagConfig.LocCodes)})";
-                }
-                else
-                {
-                    // Lenient: accept any alphanumeric string 1-8 chars, no spaces
-                    if (value.Length < 1 || value.Length > 8 || value.Contains(" ") || !value.All(char.IsLetterOrDigit))
-                        return $"LOC '{value}' must be 1-8 alphanumeric characters with no spaces";
-                }
-            }
-            else if (tokenName == ParamRegistry.ZONE)
-            {
-                // FLEX-001: Accept custom ZONE codes from config before strict/lenient check
-                if (CustomZoneCodes.Count > 0 && CustomZoneCodes.Contains(value))
-                    return null; // Custom code accepted
-                if (TagConfig.ValidateStrictMode)
-                {
-                    // F-08: Use cached HashSet for O(1) lookup instead of List.Contains O(n)
-                    var zoneSet = _zoneCodesSet ??= new HashSet<string>(TagConfig.ZoneCodes, StringComparer.OrdinalIgnoreCase);
-                    if (!zoneSet.Contains(value))
-                        return $"ZONE '{value}' not in valid set ({string.Join(",", TagConfig.ZoneCodes)})";
-                }
-                else
-                {
-                    // Lenient: accept any alphanumeric string 1-8 chars, no spaces
-                    if (value.Length < 1 || value.Length > 8 || value.Contains(" ") || !value.All(char.IsLetterOrDigit))
-                        return $"ZONE '{value}' must be 1-8 alphanumeric characters with no spaces";
-                }
-            }
-            else if (tokenName == ParamRegistry.SYS)
-            {
-                if (!ValidSysCodes.Contains(value))
-                    return $"SYS '{value}' not in valid set ({string.Join(",", ValidSysCodes)})";
-            }
-            else if (tokenName == ParamRegistry.FUNC)
-            {
-                if (!ValidFuncCodes.Contains(value))
-                    return $"FUNC '{value}' not in valid set ({string.Join(",", ValidFuncCodes)})";
-            }
-            else if (tokenName == ParamRegistry.LVL)
-            {
-                // Valid LVL codes: L00-L99, L100+, GF, LG, UG, B1-B9, B10+, SB, RF, PH, AT, TR, POD, MZ, PL
-                string lvlUpper = value.ToUpperInvariant();
-                if (lvlUpper.Length > 4 || lvlUpper.Contains(" "))
-                    return $"LVL '{value}' exceeds 4-char limit or contains spaces";
-                if (lvlUpper == "XX")
-                    return null; // XX is valid but a placeholder
-                bool isKnownLvl = lvlUpper == "GF" || lvlUpper == "RF" || lvlUpper == "LG" ||
-                    lvlUpper == "UG" || lvlUpper == "MZ" || lvlUpper == "PL" || lvlUpper == "PH" ||
-                    lvlUpper == "AT" || lvlUpper == "TR" || lvlUpper == "POD" ||
-                    (lvlUpper.StartsWith("L") && lvlUpper.Length >= 2 && lvlUpper.Length <= 4 &&
-                        lvlUpper.Substring(1).All(char.IsDigit)) ||
-                    (lvlUpper.StartsWith("B") && lvlUpper.Length >= 2 && lvlUpper.Length <= 3 &&
-                        lvlUpper.Substring(1).All(char.IsDigit)) ||
-                    (lvlUpper.StartsWith("SB") && (lvlUpper.Length == 2 ||
-                        lvlUpper.Substring(2).All(char.IsDigit)));
-                if (!isKnownLvl && !lvlUpper.All(c => char.IsLetterOrDigit(c)))
-                    return $"LVL '{value}' contains invalid characters";
-            }
-            else if (tokenName == ParamRegistry.PROD)
-            {
-                // PROD codes: 2-4 uppercase alphanumeric characters
-                if (value.Length < 2 || value.Length > 4)
-                    return $"PROD '{value}' should be 2-4 characters";
-                if (!value.All(c => char.IsLetterOrDigit(c)))
-                    return $"PROD '{value}' must be alphanumeric only";
-            }
-            else if (tokenName == ParamRegistry.SEQ)
-            {
-                if (!int.TryParse(value, out int seqVal))
-                    return $"SEQ '{value}' is not a valid number";
-                if (seqVal < 0)
-                    return $"SEQ '{value}' must be a positive number";
-                int seqWidth = TagConfig.SeqPadWidth > 0 ? TagConfig.SeqPadWidth : TagConfig.NumPad;
-                if (value.Length > seqWidth + 1)
-                    return $"SEQ '{value}' exceeds {seqWidth}-digit format";
-            }
-            return null; // valid
-        }
-
-        /// <summary>Phase 78: Validation memoization cache — caches ValidateToken results per (token,value) pair.
-        /// For 50K elements with ~200 unique token combinations, reduces validation calls from 50K×8 to ~200.
-        /// Thread-safe ConcurrentDictionary. Cleared via InvalidateValidatorCaches().</summary>
-        private static readonly System.Collections.Concurrent.ConcurrentDictionary<string, string>
-            _tokenValidationCache = new System.Collections.Concurrent.ConcurrentDictionary<string, string>();
-
-        /// <summary>Phase 78: Memoized token validation — O(1) lookup for repeated (token,value) pairs.</summary>
-        private static string ValidateTokenCached(string tokenName, string value)
-        {
-            string cacheKey = $"{tokenName}|{value ?? ""}";
-            return _tokenValidationCache.GetOrAdd(cacheKey, _ => ValidateToken(tokenName, value));
-        }
-
-        /// <summary>
-        /// Validate all 8 tokens on an element. Returns a list of validation errors
-        /// (empty list = fully valid). Uses memoized token validation for O(1) repeated lookups.
-        /// </summary>
-        // F-16: Static readonly token param array — avoids per-call allocation
-        private static readonly string[] _validateElementTokenParams = new[]
-        {
-            ParamRegistry.DISC, ParamRegistry.LOC, ParamRegistry.ZONE,
-            ParamRegistry.LVL, ParamRegistry.SYS, ParamRegistry.FUNC,
-            ParamRegistry.PROD, ParamRegistry.SEQ,
-        };
-        // F-16: [ThreadStatic] reusable error list — avoids List<ValidationError> allocation per call
-        [ThreadStatic]
-        private static List<ValidationError> _validateElementErrors;
-
-        public static List<ValidationError> ValidateElement(Element el)
-        {
-            // F-16: Reuse thread-local list instead of allocating new on each call
-            if (_validateElementErrors == null) _validateElementErrors = new List<ValidationError>();
-            else _validateElementErrors.Clear();
-            var errors = _validateElementErrors;
-            string[] tokenParams = _validateElementTokenParams;
-
-            foreach (string param in tokenParams)
-            {
-                string val = ParameterHelpers.GetString(el, param);
-                string error = ValidateTokenCached(param, val);
-                if (error != null)
-                {
-                    var errorType = string.IsNullOrEmpty(val)
-                        ? ValidationErrorType.TokenEmpty
-                        : ValidationErrorType.TokenFormat;
-                    errors.Add(new ValidationError(error, errorType));
-                }
-            }
-
-            // Cross-validate: DISC must match element category
-            // Accounts for system-aware DISC correction (pipes can be "P" when system is plumbing)
-            string catName = ParameterHelpers.GetCategoryName(el);
-            string disc = ParameterHelpers.GetString(el, ParamRegistry.DISC);
-            string sys = ParameterHelpers.GetString(el, ParamRegistry.SYS);
-            if (!string.IsNullOrEmpty(catName) && !string.IsNullOrEmpty(disc))
-            {
-                string expectedDisc = TagConfig.DiscMap.TryGetValue(catName, out string d) ? d : null;
-                // Apply system-aware DISC correction (e.g., M→P for plumbing pipes, M→FP for fire)
-                if (expectedDisc != null && !string.IsNullOrEmpty(sys))
-                    expectedDisc = TagConfig.GetSystemAwareDisc(expectedDisc, sys, catName);
-                if (expectedDisc != null && expectedDisc != disc)
-                    errors.Add(new ValidationError(
-                        $"DISC mismatch: element category '{catName}' expects '{expectedDisc}' but has '{disc}'",
-                        ValidationErrorType.CrossValidation));
-            }
-
-            // Cross-validate: SYS should be valid for this category
-            // Uses SysMap lookup to allow ALL valid SYS codes for ambiguous categories
-            // (e.g., Pipes can be DCW, DHW, SAN, RWD, GAS, HWS, FP)
-            if (!string.IsNullOrEmpty(catName) && !string.IsNullOrEmpty(sys))
-            {
-                bool sysValidForCategory = false;
-                // Check if this SYS code lists this category in SysMap
-                if (TagConfig.SysMap.TryGetValue(sys, out var sysCats) && sysCats.Contains(catName))
-                    sysValidForCategory = true;
-                // Also accept discipline-default SYS codes (ARC, STR, GEN, etc.)
-                string discForCat = TagConfig.DiscMap.TryGetValue(catName, out string dc) ? dc : "A";
-                if (sys == TagConfig.GetDiscDefaultSysCode(discForCat))
-                    sysValidForCategory = true;
-                if (!sysValidForCategory)
-                {
-                    // Find what SYS codes ARE valid for this category
-                    var validSysForCat = TagConfig.SysMap.Where(kvp => kvp.Value.Contains(catName))
-                        .Select(kvp => kvp.Key).ToList();
-                    string validList = validSysForCat.Count > 0
-                        ? string.Join("/", validSysForCat)
-                        : TagConfig.GetDiscDefaultSysCode(discForCat);
-                    errors.Add(new ValidationError(
-                        $"SYS mismatch: category '{catName}' expects '{validList}' but has '{sys}'",
-                        ValidationErrorType.CrossValidation));
-                }
-            }
-
-            // Cross-validate: PROD should be consistent with DISC/SYS
-            string prod = ParameterHelpers.GetString(el, ParamRegistry.PROD);
-            if (!string.IsNullOrEmpty(prod) && !string.IsNullOrEmpty(disc))
-            {
-                string prodError = ValidateProdForDisc(prod, disc);
-                if (prodError != null)
-                    errors.Add(new ValidationError(prodError, ValidationErrorType.CrossValidation));
-            }
-
-            // Cross-validate: FUNC should be consistent with SYS
-            // Phase 39: Expanded FUNC-SYS cross-validation with discipline-aware mapping.
-            // Each SYS code has a set of valid FUNC codes (primary + smart sub-functions).
-            // FUNC codes from unrelated disciplines are flagged as cross-validation errors.
-            string func = ParameterHelpers.GetString(el, ParamRegistry.FUNC);
-            if (!string.IsNullOrEmpty(func) && !string.IsNullOrEmpty(sys))
-            {
-                var validFuncsForSys = GetValidFuncsForSys(sys);
-                if (validFuncsForSys.Count > 0 && !validFuncsForSys.Contains(func))
-                {
-                    string expectedList = string.Join("/", validFuncsForSys);
-                    errors.Add(new ValidationError(
-                        $"FUNC '{func}' not valid for SYS '{sys}' (expected one of: {expectedList})",
-                        ValidationErrorType.CrossValidation));
-                }
-            }
-
-            // Phase 66b: Cross-validate FUNC→PROD — detect contradictory function/product pairs.
-            // e.g., FUNC=SUP (Supply) with PROD=WC (WC fixture) is contradictory.
-            if (!string.IsNullOrEmpty(func) && !string.IsNullOrEmpty(prod) &&
-                func != "GEN" && prod != "GEN")
-            {
-                string funcProdError = ValidateFuncProdPair(func, prod, disc);
-                if (funcProdError != null)
-                    errors.Add(new ValidationError(funcProdError, ValidationErrorType.CrossValidation));
-            }
-
-            // Phase 86: Return defensive copy — raw [ThreadStatic] reference would be
-            // cleared on next call, corrupting any caller that stored the result.
-            return new List<ValidationError>(errors);
-        }
-
-        /// <summary>Phase 66b: Validate FUNC→PROD pair consistency.
-        /// Detects contradictory function/product combinations like FUNC=SUP with PROD=WC.</summary>
-        // PERF: Static readonly to avoid per-call Dictionary+HashSet allocation
-        private static readonly Dictionary<string, HashSet<string>> _incompatibleFuncProdPairs =
-            new Dictionary<string, HashSet<string>>(StringComparer.OrdinalIgnoreCase)
-            {
-                // Supply function should not have sanitary/plumbing products
-                { "SUP", new HashSet<string>(StringComparer.OrdinalIgnoreCase) { "WC", "WHB", "URN", "SNK", "SHW", "BTH", "BID", "MOP" } },
-                // Return function should not have electrical products
-                { "RTN", new HashSet<string>(StringComparer.OrdinalIgnoreCase) { "DB", "MCC", "MSB", "SWB", "SKT", "LUM" } },
-                // Lighting function should not have HVAC products
-                { "LTG", new HashSet<string>(StringComparer.OrdinalIgnoreCase) { "AHU", "FCU", "VAV", "CHR", "BLR", "RAD", "DAM" } },
-                // Power function should not have plumbing products
-                { "PWR", new HashSet<string>(StringComparer.OrdinalIgnoreCase) { "WC", "WHB", "PP", "PFT", "PAC", "FPP", "TRP" } },
-                // Sanitary function should not have HVAC products
-                { "SAN", new HashSet<string>(StringComparer.OrdinalIgnoreCase) { "AHU", "FCU", "VAV", "FAN", "HRU", "DAM", "CLT" } },
-                // Fire protection function should not have architectural products
-                { "FLS", new HashSet<string>(StringComparer.OrdinalIgnoreCase) { "DR", "WIN", "WL", "FL", "CLG", "RF", "FUR" } },
-            };
-
-        private static string ValidateFuncProdPair(string func, string prod, string disc)
-        {
-            if (_incompatibleFuncProdPairs.TryGetValue(func, out var badProds) && badProds.Contains(prod))
-                return $"FUNC '{func}' is incompatible with PROD '{prod}' — check discipline assignment";
-
-            return null;
-        }
-
-        /// <summary>
-        /// Validate the format of a complete assembled tag string.
-        /// Returns null if valid, or an error description.
-        /// </summary>
-        public static string ValidateTagFormat(string tag)
-        {
-            if (string.IsNullOrEmpty(tag))
-                return "Tag is empty";
-
-            // Phase 86b: Use full separator string for multi-char separator support
-            string sepStr = !string.IsNullOrEmpty(TagConfig.Separator) ? TagConfig.Separator : "-";
-            string[] parts = tag.Split(new[] { sepStr }, StringSplitOptions.None);
-            if (parts.Length != 8)
-                return $"Tag has {parts.Length} segments (expected 8): {tag}";
-
-            for (int i = 0; i < parts.Length; i++)
-            {
-                if (string.IsNullOrWhiteSpace(parts[i]))
-                    return $"Segment {i + 1} is empty in tag: {tag}";
-            }
-
-            // Validate ALL 8 segments against their respective rules
-            // Phase 86: Use cached validation for O(1) repeated lookups
-            string discError = ValidateTokenCached(ParamRegistry.DISC, parts[0]);
-            if (discError != null) return discError;
-
-            string locError = ValidateTokenCached(ParamRegistry.LOC, parts[1]);
-            if (locError != null) return locError;
-
-            string zoneError = ValidateTokenCached(ParamRegistry.ZONE, parts[2]);
-            if (zoneError != null) return zoneError;
-
-            string lvlError = ValidateTokenCached(ParamRegistry.LVL, parts[3]);
-            if (lvlError != null) return lvlError;
-
-            string sysError = ValidateTokenCached(ParamRegistry.SYS, parts[4]);
-            if (sysError != null) return sysError;
-
-            string funcError = ValidateTokenCached(ParamRegistry.FUNC, parts[5]);
-            if (funcError != null) return funcError;
-
-            string prodError = ValidateTokenCached(ParamRegistry.PROD, parts[6]);
-            if (prodError != null) return prodError;
-
-            string seqError = ValidateTokenCached(ParamRegistry.SEQ, parts[7]);
-            if (seqError != null) return seqError;
-
-            return null; // valid
-        }
-
-        /// <summary>
-        /// Phase 39: Get valid FUNC codes for a given SYS code.
-        /// Each system has primary FUNC + allowed sub-function variants (e.g., HVAC allows SUP/RTN/EXH/FRA).
-        /// Returns empty set if SYS is unknown (no validation applied).
-        /// Cross-references CIBSE TM40, Uniclass 2015 Ss tables, and ISO 19650 annex.
-        /// </summary>
-        internal static HashSet<string> GetValidFuncsForSys(string sys)
-        {
-            EnsureValidFuncsLoaded();
-            if (_validFuncsForSys.TryGetValue(sys, out var funcs)) return funcs;
-            return new HashSet<string>(StringComparer.OrdinalIgnoreCase);
-        }
-
-        private static bool _validFuncsCsvLoaded = false;
-        private static void EnsureValidFuncsLoaded()
-        {
-            if (_validFuncsCsvLoaded) return;
-            _validFuncsCsvLoaded = true;
-            TryLoadValidFuncsFromCsv();
-        }
-
-        /// <summary>
-        /// Attempts to rebuild _validFuncsForSys from STING_FUNC_SYS_MATRIX.csv.
-        /// Falls back to the hardcoded defaults if the file is absent or malformed.
-        /// Called from LoadFromFile, LoadDefaults, and lazily from GetValidFuncsForSys.
-        /// </summary>
-        private static void TryLoadValidFuncsFromCsv()
-        {
-            try
-            {
-                string csvPath = StingToolsApp.FindDataFile("STING_FUNC_SYS_MATRIX.csv");
-                if (string.IsNullOrEmpty(csvPath) || !System.IO.File.Exists(csvPath)) return;
-
-                var loaded = new Dictionary<string, HashSet<string>>(StringComparer.OrdinalIgnoreCase);
-                bool first = true;
-                foreach (string raw in System.IO.File.ReadLines(csvPath))
-                {
-                    if (first) { first = false; continue; } // skip header
-                    string line = raw.Trim();
-                    if (string.IsNullOrEmpty(line) || line.StartsWith("#")) continue;
-
-                    // CSV: SYS_CODE,SYS_DESCRIPTION,FUNC_CODE,FUNC_DESCRIPTION,...
-                    var cols = StingToolsApp.ParseCsvLine(line);
-                    if (cols == null || cols.Length < 3) continue;
-                    string sysCode  = cols[0].Trim();
-                    string funcCode = cols[2].Trim();
-                    if (string.IsNullOrEmpty(sysCode) || string.IsNullOrEmpty(funcCode)) continue;
-
-                    if (!loaded.TryGetValue(sysCode, out var set))
-                    {
-                        set = new HashSet<string>(StringComparer.OrdinalIgnoreCase);
-                        loaded[sysCode] = set;
-                    }
-                    set.Add(funcCode);
-                }
-
-                if (loaded.Count > 0)
-                {
-                    // Merge into existing dict: CSV wins; hardcoded entries not in CSV are retained
-                    foreach (var kvp in loaded)
-                    {
-                        if (_validFuncsForSys.TryGetValue(kvp.Key, out var existing))
-                            foreach (string fc in kvp.Value) existing.Add(fc);
-                        else
-                            _validFuncsForSys[kvp.Key] = kvp.Value;
-                    }
-                    StingLog.Info($"TagConfig: merged STING_FUNC_SYS_MATRIX.csv → {loaded.Count} SYS entries into _validFuncsForSys");
-                }
-            }
-            catch (Exception ex)
-            {
-                StingLog.Warn($"TagConfig: TryLoadValidFuncsFromCsv failed — using hardcoded defaults: {ex.Message}");
-            }
-        }
-
-        /// <summary>Number of SYS entries in the valid-FUNC lookup table. Used by TagConfig.IsLoaded.</summary>
-        internal static int ValidFuncsForSysCount => _validFuncsForSys.Count;
-
-        /// <summary>Phase 39: Comprehensive SYS→FUNC mapping — hardcoded baseline.
-        /// Extended at runtime by TryLoadValidFuncsFromCsv() from STING_FUNC_SYS_MATRIX.csv.</summary>
-        private static Dictionary<string, HashSet<string>> _validFuncsForSys =
-            new Dictionary<string, HashSet<string>>(StringComparer.OrdinalIgnoreCase)
-            {
-                // Mechanical systems — aligned with STING_FUNC_SYS_MATRIX.csv
-                { "HVAC", new HashSet<string>(StringComparer.OrdinalIgnoreCase) { "SUP", "RTN", "EXH", "FRA", "HTG", "CLG", "VNT", "OA", "TRF", "REL", "GRP", "GEN" } },
-                { "HWS",  new HashSet<string>(StringComparer.OrdinalIgnoreCase) { "SUP", "RTN", "HTG", "DHW", "CND", "GEN" } },
-                { "DHW",  new HashSet<string>(StringComparer.OrdinalIgnoreCase) { "SUP", "RTN", "DHW", "GEN" } },
-                // Plumbing systems
-                { "DCW",  new HashSet<string>(StringComparer.OrdinalIgnoreCase) { "SUP", "CLD", "DCW", "GEN" } },
-                { "SAN",  new HashSet<string>(StringComparer.OrdinalIgnoreCase) { "SAN", "WAS", "SOI", "VNT", "GEN" } },
-                { "RWD",  new HashSet<string>(StringComparer.OrdinalIgnoreCase) { "RWD", "SRF", "ATT", "GEN" } },
-                { "GAS",  new HashSet<string>(StringComparer.OrdinalIgnoreCase) { "SUP", "GAS", "MED", "LOW", "GEN" } },
-                // Fire protection
-                { "FP",   new HashSet<string>(StringComparer.OrdinalIgnoreCase) { "SUP", "SYS", "WET", "DRY", "DEL", "PRE", "FOA", "FP", "FLS", "GEN" } },
-                // Electrical systems — extended for BS 7671 sub-circuits
-                { "LV",   new HashSet<string>(StringComparer.OrdinalIgnoreCase) { "PWR", "LTG", "EMG", "SML", "EAR", "UPS", "GNR", "GEN" } },
-                { "HV",   new HashSet<string>(StringComparer.OrdinalIgnoreCase) { "PWR", "TRF", "GEN" } },
-                { "FLS",  new HashSet<string>(StringComparer.OrdinalIgnoreCase) { "FLS", "GEN" } },
-                // Fire alarm
-                { "FA",   new HashSet<string>(StringComparer.OrdinalIgnoreCase) { "DET", "MCP", "SND", "AVD", "IO", "SDA", "LDS", "GEN" } },
-                // Communications / low voltage — extended for ICT/AV sub-types
-                { "COM",  new HashSet<string>(StringComparer.OrdinalIgnoreCase) { "PA", "TV", "TEL", "AV", "COM", "GEN" } },
-                { "ICT",  new HashSet<string>(StringComparer.OrdinalIgnoreCase) { "NET", "WAP", "FBR", "SER", "PTS", "ICT", "COM", "GEN" } },
-                { "NCL",  new HashSet<string>(StringComparer.OrdinalIgnoreCase) { "NCS", "EMR", "BDH", "NCL", "GEN" } },
-                { "SEC",  new HashSet<string>(StringComparer.OrdinalIgnoreCase) { "CCTV", "ACC", "INT", "DOR", "SEC", "GEN" } },
-                { "BMS",  new HashSet<string>(StringComparer.OrdinalIgnoreCase) { "MON", "CTL", "SNS", "FCT", "GEN" } },
-                // Medical gas systems (HTM 02-01)
-                { "MGS",  new HashSet<string>(StringComparer.OrdinalIgnoreCase) { "O2", "N2O", "MAP", "VAC", "EVAC", "N2", "CO2", "GEN" } },
-                // Lightning protection (BS EN 62305)
-                { "LPS",  new HashSet<string>(StringComparer.OrdinalIgnoreCase) { "AIR", "DOW", "ERT", "BND", "SPD", "TST", "GEN" } },
-                // Radiation protection (NCRP 147)
-                { "RAD",  new HashSet<string>(StringComparer.OrdinalIgnoreCase) { "SHD", "ZNE", "MON", "GEN" } },
-                // Architectural / structural / general
-                { "ARC",  new HashSet<string>(StringComparer.OrdinalIgnoreCase) { "WLL", "FLR", "CLG", "ROF", "DOR", "WIN", "STR", "RMP", "FAS", "FIT", "GEN" } },
-                { "STR",  new HashSet<string>(StringComparer.OrdinalIgnoreCase) { "COL", "BM", "SLB", "FND", "WLL", "BRC", "TRS", "RBR", "STR", "GEN" } },
-                { "GEN",  new HashSet<string>(StringComparer.OrdinalIgnoreCase) { "GEN" } },
-            };
-
-        /// <summary>Known PROD codes by discipline group for cross-validation.</summary>
-        private static readonly Dictionary<string, HashSet<string>> ProdCodesByDisc =
-            new Dictionary<string, HashSet<string>>(StringComparer.OrdinalIgnoreCase)
-            {
-                { "M", new HashSet<string>(StringComparer.OrdinalIgnoreCase) {
-                    "AHU", "FCU", "VAV", "CHR", "BLR", "PMP", "FAN", "HRU", "SPL", "GRL",
-                    "DAC", "DFT", "DU", "FDU", "IND", "RAD", "DAM", "CLT", "VFD", "SLV", "GEN" } },
-                { "E", new HashSet<string>(StringComparer.OrdinalIgnoreCase) {
-                    "DB", "MCC", "MSB", "SWB", "UPS", "TRF", "ATS", "VFD", "SPD", "RCD",
-                    "ISO", "SFS", "BKP", "SKT", "LUM", "LDV", "DWN", "LIN", "SPT", "WSH",
-                    "BOL", "UPL", "FLD", "EML", "TRK", "DEC", "CDT", "CFT", "CBLT", "CTF", "SLV", "GEN" } },
-                { "P", new HashSet<string>(StringComparer.OrdinalIgnoreCase) {
-                    "PP", "PFT", "PAC", "FPP", "FIX", "WC", "WHB", "URN", "SNK", "SHW",
-                    "BTH", "DRK", "CWL", "TRP", "BID", "EWS", "MOP", "SLV", "GEN" } },
-                { "FP", new HashSet<string>(StringComparer.OrdinalIgnoreCase) {
-                    "SPR", "FAD", "SML", "MCP", "BLL", "STB", "HTD", "FIM", "PP", "PFT",
-                    "PAC", "SLV", "GEN" } },
-                { "A", new HashSet<string>(StringComparer.OrdinalIgnoreCase) {
-                    "DR", "WIN", "WL", "FL", "CLG", "RF", "RM", "FUR", "FUS", "CWK",
-                    "RLG", "STR", "RMP", "CPN", "MUL", "CWS", "GEN" } },
-                { "S", new HashSet<string>(StringComparer.OrdinalIgnoreCase) {
-                    "COL", "BM", "FDN", "GEN" } },
-                { "LV", new HashSet<string>(StringComparer.OrdinalIgnoreCase) {
-                    "COM", "DAT", "NCL", "SEC", "TEL", "DB", "SKT", "LUM", "LDV",
-                    "CDT", "CFT", "CBLT", "CTF", "GEN" } },
-            };
-
-        /// <summary>Static lookup for DISC → valid SYS codes. Used in ValidateCrossSystemConsistency.</summary>
-        internal static readonly Dictionary<string, HashSet<string>> _validSysForDisc =
-            new Dictionary<string, HashSet<string>>
-            {
-                { "M",  new HashSet<string> { "HVAC", "HWS", "DCW", "DHW", "GAS", "RWD", "SAN" } },
-                { "E",  new HashSet<string> { "LV", "FLS", "SEC", "ICT", "COM", "NCL" } },
-                { "P",  new HashSet<string> { "DCW", "DHW", "SAN", "RWD", "GAS" } },
-                { "FP", new HashSet<string> { "FP", "FLS" } },
-                { "A",  new HashSet<string> { "ARC" } },
-                { "S",  new HashSet<string> { "STR" } },
-                { "LV", new HashSet<string> { "LV", "ICT", "COM", "SEC", "NCL" } },
-            };
-
-        /// <summary>
-        /// Validate that a PROD code is reasonable for the given DISC.
-        /// Returns null if valid, error message if mismatched.
-        /// Only flags clear mismatches (e.g., plumbing PROD on an M element).
-        /// </summary>
-        private static string ValidateProdForDisc(string prod, string disc)
-        {
-            // GEN is valid for all disciplines
-            if (prod == "GEN" || prod == "SPE" || prod == "MED") return null;
-            // If we don't have a mapping for this discipline, skip
-            if (!ProdCodesByDisc.TryGetValue(disc, out var validProds)) return null;
-            // VFD appears in both M and E — skip cross-disc check for shared codes
-            if (prod == "VFD" || prod == "PMP") return null;
-            // Only flag if the PROD is known to belong exclusively to a different discipline
-            if (!validProds.Contains(prod))
-            {
-                // Check if the PROD belongs to a clearly different discipline
-                foreach (var kvp in ProdCodesByDisc)
-                {
-                    if (kvp.Key != disc && kvp.Value.Contains(prod))
-                        return $"PROD '{prod}' typically belongs to DISC '{kvp.Key}', not '{disc}'";
-                }
-            }
-            return null;
-        }
-    }
 
     /// <summary>
     /// Ported from tag_config.py — project-level ISO 19650 token lookup tables.
@@ -7343,535 +6424,4 @@ namespace StingTools.Core
         }
     }
 
-    /// <summary>
-    /// Advanced tagging intelligence engine providing multi-layered reasoning,
-    /// cross-validation, workset inference, connected system traversal,
-    /// sizing-based classification, and confidence scoring.
-    ///
-    /// Intelligence layers (applied in order of reliability):
-    ///   L1. Connected MEP system name (connector traversal)
-    ///   L2. Duct/Pipe system type built-in parameter
-    ///   L3. Electrical circuit panel analysis
-    ///   L4. Family name pattern matching (35+ equipment types)
-    ///   L5. Room-type inference (Server Room → ICT, Kitchen → SAN)
-    ///   L6. Workset name inference (M-Mechanical → M, E-Electrical → E)
-    ///   L7. Connected element traversal (trace pipe/duct to source equipment)
-    ///   L8. Size-based classification (small pipe → sanitary, large pipe → DCW)
-    ///   L9. Adjacent element analysis (nearby elements suggest system context)
-    ///   L10. Cross-validation (DISC vs SYS vs FUNC consistency checks)
-    ///
-    /// Each layer produces a confidence score (0.0-1.0). The highest-confidence
-    /// result wins. Ties are broken by layer priority (lower = more reliable).
-    /// </summary>
-    public static class TagIntelligence
-    {
-        /// <summary>
-        /// Result from an intelligence layer, including the derived value,
-        /// the confidence level, and which layer produced it (for audit trail).
-        /// </summary>
-        public class InferenceResult
-        {
-            public string Value { get; set; }
-            public double Confidence { get; set; }
-            public string Source { get; set; }
-
-            public InferenceResult(string value, double confidence, string source)
-            {
-                Value = value;
-                Confidence = confidence;
-                Source = source;
-            }
-        }
-
-        // ── Layer 6: Workset Name Inference ──────────────────────────────────
-
-        /// <summary>
-        /// Infer discipline from the element's workset name.
-        /// Revit worksets follow naming conventions like "M-Mechanical", "E-Electrical",
-        /// "P-Plumbing", "A-Architecture", "S-Structure" per AEC UK BIM Protocol.
-        /// </summary>
-        public static InferenceResult InferDiscFromWorkset(Element el)
-        {
-            try
-            {
-                if (el.Document.IsWorkshared)
-                {
-                    WorksetId wsId = el.WorksetId;
-                    if (wsId != null && wsId != WorksetId.InvalidWorksetId)
-                    {
-                        WorksetTable table = el.Document.GetWorksetTable();
-                        Workset ws = table.GetWorkset(wsId);
-                        if (ws != null)
-                        {
-                            string wsName = ws.Name?.ToUpperInvariant() ?? "";
-
-                            if (wsName.StartsWith("M-") || wsName.Contains("MECHANICAL") ||
-                                wsName.Contains("HVAC"))
-                                return new InferenceResult("M", 0.7, "Workset: " + ws.Name);
-                            if (wsName.StartsWith("E-") || wsName.Contains("ELECTRICAL") ||
-                                wsName.Contains("LIGHTING"))
-                                return new InferenceResult("E", 0.7, "Workset: " + ws.Name);
-                            if (wsName.StartsWith("P-") || wsName.Contains("PLUMBING") ||
-                                wsName.Contains("PUBLIC HEALTH"))
-                                return new InferenceResult("P", 0.7, "Workset: " + ws.Name);
-                            if (wsName.StartsWith("A-") || wsName.Contains("ARCHITECT"))
-                                return new InferenceResult("A", 0.7, "Workset: " + ws.Name);
-                            if (wsName.StartsWith("S-") || wsName.Contains("STRUCT"))
-                                return new InferenceResult("S", 0.7, "Workset: " + ws.Name);
-                            if (wsName.Contains("FIRE"))
-                                return new InferenceResult("FP", 0.7, "Workset: " + ws.Name);
-                        }
-                    }
-                }
-            }
-            catch (Exception ex)
-            {
-                StingLog.Warn($"InferDiscFromWorkset: {ex.Message}");
-            }
-            return null;
-        }
-
-        // ── Layer 7: Connected Element Traversal ─────────────────────────────
-
-        /// <summary>
-        /// Trace the connected system from an element back to its source equipment.
-        /// For pipes/ducts, follows the system to find the connected major equipment
-        /// (AHU, pump, boiler, chiller) which identifies the system type definitively.
-        /// Limited to 2 hops to avoid performance issues.
-        /// </summary>
-        public static InferenceResult InferSysFromConnectedEquipment(Element el)
-        {
-            try
-            {
-                FamilyInstance fi = el as FamilyInstance;
-                if (fi?.MEPModel?.ConnectorManager == null) return null;
-
-                // Traverse up to 2 hops of connected elements
-                var visited = new HashSet<ElementId> { el.Id };
-                var queue = new Queue<(Element elem, int depth)>();
-
-                foreach (Connector conn in fi.MEPModel.ConnectorManager.Connectors)
-                {
-                    if (conn.AllRefs == null) continue;
-                    foreach (Connector other in conn.AllRefs)
-                    {
-                        if (other.Owner != null && !visited.Contains(other.Owner.Id))
-                        {
-                            visited.Add(other.Owner.Id);
-                            queue.Enqueue((other.Owner, 1));
-                        }
-                    }
-                }
-
-                while (queue.Count > 0)
-                {
-                    var (current, depth) = queue.Dequeue();
-                    string catName = ParameterHelpers.GetCategoryName(current);
-
-                    // If we reached major equipment, use its classification
-                    if (catName == "Mechanical Equipment")
-                    {
-                        string famName = ParameterHelpers.GetFamilyName(current).ToUpperInvariant();
-                        if (famName.Contains("AHU") || famName.Contains("AIR HANDLING"))
-                            return new InferenceResult("HVAC", 0.9, "Connected to AHU: " + current.Id);
-                        if (famName.Contains("BOILER") || famName.Contains("BLR"))
-                            return new InferenceResult("HWS", 0.9, "Connected to boiler: " + current.Id);
-                        if (famName.Contains("CHILLER") || famName.Contains("CHR"))
-                            return new InferenceResult("HVAC", 0.9, "Connected to chiller: " + current.Id);
-                        if (famName.Contains("PUMP"))
-                            return new InferenceResult("DCW", 0.8, "Connected to pump: " + current.Id);
-                    }
-                    else if (catName == "Electrical Equipment")
-                    {
-                        return new InferenceResult("LV", 0.85, "Connected to panel: " + current.Id);
-                    }
-
-                    // Continue traversal (max 2 hops)
-                    if (depth < 2 && current is FamilyInstance fi2 &&
-                        fi2.MEPModel?.ConnectorManager != null)
-                    {
-                        foreach (Connector conn in fi2.MEPModel.ConnectorManager.Connectors)
-                        {
-                            if (conn.AllRefs == null) continue;
-                            foreach (Connector other in conn.AllRefs)
-                            {
-                                if (other.Owner != null && !visited.Contains(other.Owner.Id))
-                                {
-                                    visited.Add(other.Owner.Id);
-                                    queue.Enqueue((other.Owner, depth + 1));
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-            catch (Exception ex)
-            {
-                StingLog.Warn($"InferSysFromConnectedEquipment: {ex.Message}");
-            }
-            return null;
-        }
-
-        // ── Layer 8: Size-Based Classification ───────────────────────────────
-
-        /// <summary>
-        /// Infer system type from element sizing.
-        /// In MEP design, pipe/duct sizes correlate strongly with system type:
-        ///   - Pipes ≤32mm: typically sanitary branch, DCW branch
-        ///   - Pipes 40-80mm: sanitary mains, DCW mains
-        ///   - Pipes 80-200mm: fire mains, DCW risers, HVAC CHW
-        ///   - Pipes ≥200mm: fire protection mains, DHW risers
-        ///   - Ducts ≤300mm: extract/exhaust branches
-        ///   - Ducts 300-600mm: supply/return branches
-        ///   - Ducts ≥600mm: supply/return mains
-        /// </summary>
-        public static InferenceResult InferSysFromSize(Element el)
-        {
-            try
-            {
-                // Read calculated size or diameter
-                Parameter sizePar = el.get_Parameter(BuiltInParameter.RBS_CALCULATED_SIZE);
-                Parameter diaPar = el.get_Parameter(BuiltInParameter.RBS_PIPE_DIAMETER_PARAM);
-
-                double diameterMm = 0;
-                if (diaPar != null && diaPar.HasValue && diaPar.StorageType == StorageType.Double)
-                    diameterMm = diaPar.AsDouble() * 304.8; // feet to mm
-
-                string catName = ParameterHelpers.GetCategoryName(el);
-
-                // Pipe size-based inference (only applies if no system info available)
-                if (catName == "Pipes" && diameterMm > 0)
-                {
-                    if (diameterMm >= 100)
-                        return new InferenceResult("FP", 0.4,
-                            $"Size inference: pipe {diameterMm:F0}mm ≥ 100mm (possible fire main)");
-                    if (diameterMm <= 32)
-                        return new InferenceResult("SAN", 0.3,
-                            $"Size inference: pipe {diameterMm:F0}mm ≤ 32mm (possible sanitary branch)");
-                }
-            }
-            catch (Exception ex)
-            {
-                StingLog.Warn($"InferSysFromSize: {ex.Message}");
-            }
-            return null;
-        }
-
-        // ── Layer 9: Adjacent Element SYS Inference ──────────────────────────
-
-        /// <summary>
-        /// ENH-004: Infer SYS from adjacent elements within a 500mm radius.
-        /// Uses BoundingBoxIntersectsFilter to find nearby elements with confirmed SYS values.
-        /// If 80%+ of adjacent elements agree on a SYS code, returns that code with confidence 0.3.
-        /// Useful for unconnected light fittings, generic models, and equipment placed near systems.
-        /// </summary>
-        public static InferenceResult InferSysFromAdjacentElements(Document doc, Element el)
-        {
-            try
-            {
-                BoundingBoxXYZ bb = el.get_BoundingBox(null);
-                if (bb == null) return null;
-
-                // Expand bounding box by 500mm (≈1.64 ft) in all directions
-                double expandFt = 500.0 / 304.8; // mm to feet
-                XYZ min = new XYZ(bb.Min.X - expandFt, bb.Min.Y - expandFt, bb.Min.Z - expandFt);
-                XYZ max = new XYZ(bb.Max.X + expandFt, bb.Max.Y + expandFt, bb.Max.Z + expandFt);
-                Outline outline = new Outline(min, max);
-
-                var bbFilter = new BoundingBoxIntersectsFilter(outline);
-                var nearby = new FilteredElementCollector(doc)
-                    .WhereElementIsNotElementType()
-                    .WherePasses(bbFilter)
-                    .Where(e => e.Id != el.Id && e.Category != null)
-                    .ToList();
-
-                if (nearby.Count == 0) return null;
-
-                // Count SYS values on nearby elements
-                var sysCounts = new Dictionary<string, int>(StringComparer.OrdinalIgnoreCase);
-                int withSys = 0;
-
-                foreach (Element adj in nearby)
-                {
-                    string adjSys = ParameterHelpers.GetString(adj, ParamRegistry.SYS);
-                    if (string.IsNullOrEmpty(adjSys)) continue;
-
-                    withSys++;
-                    sysCounts.TryGetValue(adjSys, out int sc);
-                    sysCounts[adjSys] = sc + 1;
-                }
-
-                // R3-FIX-01: Guard against empty sysCounts before .First()
-                if (withSys < 2 || sysCounts.Count == 0) return null; // Need at least 2 neighbours with SYS
-
-                // Find dominant SYS
-                var dominant = sysCounts.OrderByDescending(x => x.Value).First();
-                double agreement = (double)dominant.Value / withSys;
-
-                if (agreement >= 0.8)
-                {
-                    return new InferenceResult(dominant.Key, 0.3,
-                        $"Adjacent element inference: {dominant.Value}/{withSys} neighbours have SYS={dominant.Key}");
-                }
-            }
-            catch (Exception ex)
-            {
-                StingLog.Warn($"InferSysFromAdjacentElements: {ex.Message}");
-            }
-            return null;
-        }
-
-        // ── Layer 10: Cross-Validation ───────────────────────────────────────
-
-        /// <summary>
-        /// Cross-validate DISC, SYS, FUNC, and PROD codes against each other.
-        /// Returns a list of inconsistencies found (empty list = all valid).
-        ///
-        /// Rules:
-        ///   - DISC=M requires SYS in {HVAC, HWS, DCW, GAS, RWD, SAN, DHW}
-        ///   - DISC=E requires SYS in {LV, FLS, SEC, ICT, COM, NCL}
-        ///   - DISC=P requires SYS in {DCW, DHW, SAN, RWD, GAS}
-        ///   - DISC=FP requires SYS in {FP, FLS}
-        ///   - PROD code must be compatible with category
-        ///   - FUNC code must be compatible with SYS code
-        /// </summary>
-        public static List<string> CrossValidate(Element el)
-        {
-            var issues = new List<string>();
-            string disc = ParameterHelpers.GetString(el, ParamRegistry.DISC);
-            string sys = ParameterHelpers.GetString(el, ParamRegistry.SYS);
-            string func = ParameterHelpers.GetString(el, ParamRegistry.FUNC);
-            string prod = ParameterHelpers.GetString(el, ParamRegistry.PROD);
-            string catName = ParameterHelpers.GetCategoryName(el);
-
-            if (string.IsNullOrEmpty(disc)) return issues;
-
-            // DISC ↔ SYS consistency
-            if (!string.IsNullOrEmpty(sys) && ISO19650Validator._validSysForDisc.TryGetValue(disc, out var validSys))
-            {
-                if (!validSys.Contains(sys))
-                    issues.Add($"DISC={disc} incompatible with SYS={sys} (expected: {string.Join("/", validSys)})");
-            }
-
-            // DISC ↔ Category consistency
-            string expectedDisc = TagConfig.DiscMap.TryGetValue(catName, out string ed) ? ed : null;
-            if (expectedDisc != null && expectedDisc != disc)
-                issues.Add($"DISC={disc} doesn't match category '{catName}' (expected: {expectedDisc})");
-
-            // SYS ↔ FUNC consistency (HVAC should have SUP/RTN/EXH/FRA, not PWR)
-            if (sys == "HVAC" && func == "PWR")
-                issues.Add($"SYS=HVAC with FUNC=PWR is invalid (expected: SUP/RTN/EXH/FRA)");
-            if (sys == "LV" && (func == "SUP" || func == "HTG"))
-                issues.Add($"SYS=LV with FUNC={func} is invalid (expected: PWR/LTG)");
-
-            return issues;
-        }
-
-        /// <summary>
-        /// Perform full tagging intelligence analysis on an element.
-        /// Runs all inference layers and returns a consolidated report
-        /// with the best result per token and confidence scores.
-        /// Used by AutoPopulate and pre-tagging audit.
-        /// </summary>
-        public static Dictionary<string, InferenceResult> AnalyzeElement(Document doc, Element el)
-        {
-            var results = new Dictionary<string, InferenceResult>();
-            string catName = ParameterHelpers.GetCategoryName(el);
-
-            // DISC — category is primary (confidence 1.0), workset as validation
-            if (TagConfig.DiscMap.TryGetValue(catName, out string disc))
-                results["DISC"] = new InferenceResult(disc, 1.0, "Category: " + catName);
-
-            var wsResult = InferDiscFromWorkset(el);
-            if (wsResult != null && results.TryGetValue("DISC", out var discResult) && discResult.Value != wsResult.Value)
-                StingLog.Warn($"Element {el.Id}: category says DISC={discResult.Value} but workset says {wsResult.Value}");
-
-            // SYS — multi-layer with confidence scoring
-            string sys = TagConfig.GetMepSystemAwareSysCode(el, catName);
-            if (!string.IsNullOrEmpty(sys))
-                results["SYS"] = new InferenceResult(sys, 0.85, "MEP system detection");
-
-            // Try connected equipment traversal for higher confidence
-            var connResult = InferSysFromConnectedEquipment(el);
-            if (connResult != null && connResult.Confidence > (results.TryGetValue("SYS", out var curSys) ? curSys.Confidence : 0))
-                results["SYS"] = connResult;
-
-            // Size-based only if nothing else worked
-            if (!results.TryGetValue("SYS", out var sysEntry) || sysEntry.Confidence < 0.5)
-            {
-                var sizeResult = InferSysFromSize(el);
-                if (sizeResult != null)
-                    results["SYS"] = sizeResult;
-            }
-
-            // ENH-004: Layer 9 — adjacent element inference (lowest confidence, last resort)
-            if (!results.TryGetValue("SYS", out sysEntry) || sysEntry.Confidence < 0.3)
-            {
-                var adjResult = InferSysFromAdjacentElements(doc, el);
-                if (adjResult != null)
-                    results["SYS"] = adjResult;
-            }
-
-            // FUNC — smart detection
-            string sysVal = results.TryGetValue("SYS", out var sysForFunc) ? sysForFunc.Value : "";
-            string func = TagConfig.GetSmartFuncCode(el, sysVal);
-            if (!string.IsNullOrEmpty(func))
-                results["FUNC"] = new InferenceResult(func, 0.8, "Smart FUNC detection");
-
-            // PROD — family-aware
-            string prod = TagConfig.GetFamilyAwareProdCode(el, catName);
-            if (!string.IsNullOrEmpty(prod))
-                results["PROD"] = new InferenceResult(prod, 0.9, "Family-aware PROD");
-
-            // LVL
-            string lvl = ParameterHelpers.GetLevelCode(doc, el);
-            if (lvl != "XX")
-                results["LVL"] = new InferenceResult(lvl, 1.0, "Level: auto-derived");
-
-            return results;
-        }
-
-        /// <summary>
-        /// Generate a human-readable audit trail for an element's tag derivation.
-        /// Shows what each intelligence layer detected and why, with confidence scores.
-        /// </summary>
-        public static string GenerateAuditTrail(Document doc, Element el)
-        {
-            var sb = new System.Text.StringBuilder();
-            string catName = ParameterHelpers.GetCategoryName(el);
-            sb.AppendLine($"Element {el.Id} [{catName}]: {ParameterHelpers.GetFamilyName(el)}");
-
-            var analysis = AnalyzeElement(doc, el);
-            foreach (var kvp in analysis)
-            {
-                sb.AppendLine($"  {kvp.Key} = {kvp.Value.Value} " +
-                    $"(confidence: {kvp.Value.Confidence:P0}, source: {kvp.Value.Source})");
-            }
-
-            // Cross-validation
-            var issues = CrossValidate(el);
-            if (issues.Count > 0)
-            {
-                sb.AppendLine("  WARNINGS:");
-                foreach (string issue in issues)
-                    sb.AppendLine($"    ! {issue}");
-            }
-            else
-            {
-                sb.AppendLine("  Cross-validation: PASS");
-            }
-
-            return sb.ToString();
-        }
-    }
-
-    /// <summary>
-    /// Plugin hook system — extensibility framework for third-party command registration.
-    /// Third-party plugins register hooks at Revit startup; STING invokes them at defined points.
-    /// </summary>
-    public static class StingPluginHooks
-    {
-        /// <summary>Hook invoked before each element is tagged in RunFullPipeline.</summary>
-        public static event Action<Document, Element> BeforeTagElement;
-
-        /// <summary>Hook invoked after each element is tagged in RunFullPipeline.</summary>
-        public static event Action<Document, Element, string> AfterTagElement;
-
-        /// <summary>Hook for custom token validation. Return null if valid, error string if invalid.</summary>
-        public static event Func<string, string, string> ValidateToken;
-
-        /// <summary>Hook invoked before a workflow preset executes.</summary>
-        public static event Action<string> BeforeWorkflow;
-
-        /// <summary>Hook invoked after a workflow preset completes.</summary>
-        public static event Action<string, bool> AfterWorkflow;
-
-        /// <summary>Registry of third-party commands keyed by tag string.</summary>
-        private static readonly Dictionary<string, Func<UIApplication, string>> _customCommands
-            = new Dictionary<string, Func<UIApplication, string>>(StringComparer.OrdinalIgnoreCase);
-
-        /// <summary>Invoke BeforeWorkflow event (callable from other classes since events can only be raised by declaring class).</summary>
-        internal static void InvokeBeforeWorkflow(string presetName) => BeforeWorkflow?.Invoke(presetName);
-
-        /// <summary>Invoke AfterWorkflow event (callable from other classes since events can only be raised by declaring class).</summary>
-        internal static void InvokeAfterWorkflow(string presetName, bool success) => AfterWorkflow?.Invoke(presetName, success);
-
-        /// <summary>Register a custom command that can be invoked from workflows or dispatch.</summary>
-        public static void RegisterCommand(string tag, Func<UIApplication, string> handler)
-        {
-            if (string.IsNullOrWhiteSpace(tag) || handler == null) return;
-            _customCommands[tag] = handler;
-            StingLog.Info($"StingPluginHooks: registered custom command '{tag}'");
-        }
-
-        /// <summary>Unregister a custom command.</summary>
-        public static void UnregisterCommand(string tag)
-        {
-            if (_customCommands.Remove(tag))
-                StingLog.Info($"StingPluginHooks: unregistered command '{tag}'");
-        }
-
-        /// <summary>Try to execute a registered custom command. Returns (found, resultMessage).</summary>
-        public static (bool Found, string Result) TryExecuteCommand(string tag, UIApplication app)
-        {
-            if (_customCommands.TryGetValue(tag, out var handler))
-            {
-                try
-                {
-                    string result = handler(app);
-                    return (true, result ?? "OK");
-                }
-                catch (Exception ex)
-                {
-                    StingLog.Error($"StingPluginHooks: command '{tag}' failed", ex);
-                    return (true, $"Error: {ex.Message}");
-                }
-            }
-            return (false, null);
-        }
-
-        /// <summary>Get list of registered custom command tags.</summary>
-        public static IReadOnlyList<string> RegisteredCommands => _customCommands.Keys.ToList().AsReadOnly();
-
-        /// <summary>Fire the BeforeTagElement hook (safe — catches exceptions).</summary>
-        internal static void FireBeforeTag(Document doc, Element el)
-        {
-            try { BeforeTagElement?.Invoke(doc, el); }
-            catch (Exception ex) { StingLog.Warn($"StingPluginHooks.BeforeTag: {ex.Message}"); }
-        }
-
-        /// <summary>Fire the AfterTagElement hook (safe — catches exceptions).</summary>
-        internal static void FireAfterTag(Document doc, Element el, string tag)
-        {
-            try { AfterTagElement?.Invoke(doc, el, tag); }
-            catch (Exception ex) { StingLog.Warn($"StingPluginHooks.AfterTag: {ex.Message}"); }
-        }
-
-        /// <summary>Run custom validators. Returns first error or null.</summary>
-        internal static string RunCustomValidators(string tokenName, string value)
-        {
-            if (ValidateToken == null) return null;
-            foreach (var handler in ValidateToken.GetInvocationList().Cast<Func<string, string, string>>())
-            {
-                try
-                {
-                    string error = handler(tokenName, value);
-                    if (error != null) return error;
-                }
-                catch (Exception ex) { StingLog.Warn($"StingPluginHooks.ValidateToken: {ex.Message}"); }
-            }
-            return null;
-        }
-
-        /// <summary>Clear all hooks and registered commands (called on plugin shutdown).</summary>
-        public static void ClearAll()
-        {
-            BeforeTagElement = null;
-            AfterTagElement = null;
-            ValidateToken = null;
-            BeforeWorkflow = null;
-            AfterWorkflow = null;
-            _customCommands.Clear();
-        }
-    }
 }
